+++ src/components/Cataloguing/Marc/Record/index.js
@@ -7,16 +7,13 @@ import {
   AccordionSet,
   Callout,
   Row,
-  PaneMenu,
   Button,
   Col,
   Icon,
-  EmptyMessage,
   MenuSection,
   PaneFooter
 } from '@folio/stripes/components';
 import { AppIcon } from '@folio/stripes-core';
-import { FormattedMessage } from 'react-intl';
 import { bindActionCreators } from 'redux';
 import { union, sortBy, includes } from 'lodash';
 import {
@@ -51,7 +48,6 @@ import { formFieldValue, resolve } from '../../../../redux/helpers/Selector';
 import { TAGS, TAG_NOT_REPEATABLE } from '../../Utils/MarcConstant';
 import DataFieldForm from '../Form/DataField';
 import VariableFieldForm from '../Form/VariableField';
-import { validateTag } from '../../../../shared/utils/Function';
 
 class Record extends React.Component<
   Props,
@@ -368,7 +364,7 @@ class Record extends React.Component<
   };
 
   handleClose = (checkCodeForSearch, idFromBibRecord) => {
-    const { id, submit } = this.state;
+    const { submit } = this.state;
     const { dispatch, router, toggleFilterPane, reset } = this.props;
     if (checkCodeForSearch === 'OK') {
       dispatch({