+++ test/bigtest/.eslintrc.js
@@ -1,9 +1,7 @@
 module.exports = {
-    rules: {
-      "func-names": "off",
-      "no-unused-expressions": "off",
-      "one-var": "off",
-      "no-undef": "off"
-    }
-  };
-  
+  rules: {
+    "func-names": "off",
+    "no-unused-expressions": "off",
+    "one-var": "off"
+  }
+};