+++ test/bigtest/helpers/setup-application.js
@@ -1,15 +1,28 @@
 import setupStripesCore from '@folio/stripes-core/test/bigtest/helpers/setup-application';
 import mirageOptions from '../network';
 
+mirageOptions.serverType = 'miragejs';
+
 export default function setupApplication({
   scenarios,
-  hasAllPerms = true
+  permissions = {},
+  hasAllPerms = true,
+  modules,
+  translations,
+  currentUser,
 } = {}) {
   setupStripesCore({
-    mirageOptions,
+    mirageOptions: {
+      serverType: 'miragejs',
+      ...mirageOptions
+    },
     scenarios,
+    permissions,
+    modules,
+    translations,
     stripesConfig: {
-      hasAllPerms
-    }
+      hasAllPerms,
+    },
+    currentUser,
   });
 }