+++ test/bigtest/interactors/marccat.js
@@ -1,5 +1,4 @@
 import { interactor } from '@bigtest/interactor';
 
 export default @interactor class MarccatInteractor {
-  static defaultScope = '#ModuleContainer';
 }