+++ test/bigtest/network/index.js
@@ -1,4 +1,4 @@
-import { camelize } from '@bigtest/mirage';
+import { camelize, underscore } from 'inflected';
 
 // auto-import all mirage submodules
 const req = require.context('./', true, /\.js$/);
@@ -8,7 +8,7 @@ const modules = req.keys().reduce((acc, modulePath) => {
   const moduleName = moduleParts[2];
 
   if (moduleName) {
-    const moduleKey = camelize(moduleName.replace('.js', ''));
+    const moduleKey = camelize(underscore(moduleName.replace('.js', '')), false);
 
     return Object.assign(acc, {
       [moduleType]: {
@@ -20,9 +20,9 @@ const modules = req.keys().reduce((acc, modulePath) => {
     return Object.assign(acc, {
       baseConfig: req(modulePath).default
     });
-  } else {
-    return acc;
   }
+
+  return acc;
 }, {});
 
 export default modules;