+++ test/bigtest/network/scenarios/default.js
@@ -1,5 +1,7 @@
 /* istanbul ignore file */
 
 // default scenario is used during `yarn start --mirage`
-export default function defaultScenario() {
+export default function defaultScenario(server) {
+  server.create('fromTemplate', 1);
+  server.createList('mergedSearch', 1);
 }